feat: rename session routes, add <SessionView> component (work in progress) to manage sessions
<SessionView> right now renders a progress bar of the session, a flashcard, an input field and a
submit button. Allows the user to submit an answer for the shown flashcard, after answering
increases progress and shows back of the card, then user can click on next. After 'next' click, uses
fun generator function to move to the next card and shows the front of it to repeat the process.
Work in progress.

re #3