feat: refactor for better maintainability and code quality, add improved finished state
handling the UI states using MODES worked but it was too limiting to only use one string that
alternated between modes. That works if the UI depends on one factor but the slightly increased
complexity of the UI depended on multiple factors/variables, what action the user could take at the
current moment (ANSWER, START_NEXT, FINISH), and what was the result of the user's last action
(correct or incorrect answer). technically more but those are the main ones. Refactored to use a
reducer to always have more data about the ui state and for the added abstraction. Using an object
with useState could've been viable too. Also, added a "Finish!" button label and made it so that
that button in that state returns the user to the dashboard