feat: implement proof of concept to fetch data for the dashboard from database
add a fetch function that runs in the server w/a db query that returns data for decks. including
lessons count and reviews count for the cards. (joins decks and cards table and uses subqueries).
restructure dashboard page to run the quer in the backend. render each <Deck /> individually rather
than rendering a <Decks /> component.